+++ .gitignore
@@ -4,18 +4,19 @@ bin
 # cmake build directory
 build
 .lintbin
-nvfuser/version.py
-nvfuser/include
-nvfuser/lib
-nvfuser/share
-nvfuser/cmake
 
-python/build
-python/nvfuser/version.py
-python/nvfuser/include
-python/nvfuser/lib
-python/nvfuser/share
-python/nvfuser/cmake
+nvfuser_common/version.py
+nvfuser_common/include
+nvfuser_comon/lib
+nvfuser_common/share
+nvfuser_common/cmake
+
+python/nvfuser_common/build
+python/nvfuser_common/version.py
+python/nvfuser_common/include
+python/nvfuser_common/lib
+python/nvfuser_common/share
+python/nvfuser_common/cmake
 
 .hypothesis
 *.egg-info/
