+++ examples/sinh_libtorch/README.md
@@ -9,7 +9,7 @@
 ```
 mkdir build
 cd build
-CMAKE_PREFIX_PATH=$(python -c 'import nvfuser.utils; import torch.utils; print(nvfuser.utils.cmake_prefix_path, torch.utils.cmake_prefix_path, sep=";")')
+CMAKE_PREFIX_PATH=$(python -c 'import nvfuser_common.utils; import torch.utils; print(nvfuser_common.utils.cmake_prefix_path, torch.utils.cmake_prefix_path, sep=";")')
 cmake -DCMAKE_PREFIX_PATH=$CMAKE_PREFIX_PATH -G Ninja ..
 cmake --build .
 ```
