+++ python/tools/gen_nvfuser_version.py
@@ -70,6 +70,6 @@ def get_pytorch_use_distributed():
 
 
 if __name__ == "__main__":
-    version_file = nvfuser_root / "nvfuser" / "version.py"
+    version_file = nvfuser_root / "nvfuser_common" / "version.py"
     with open(version_file, "w") as f:
         f.write("_version_str = '{}'\n".format(get_version()))
