+++ python/utils.py
@@ -285,7 +285,9 @@ class build_ext(setuptools.command.build_ext.build_ext):
             filename = self.get_ext_filename(self.get_ext_fullname(ext.name))
             fileext = os.path.splitext(filename)[1]
 
-            libnvfuser_path = os.path.join("./nvfuser/lib", f"libnvfuser{fileext}")
+            libnvfuser_path = os.path.join(
+                "./nvfuser_common/lib", f"libnvfuser{fileext}"
+            )
             assert os.path.exists(libnvfuser_path)
             install_dst = os.path.join(self.build_lib, filename)
             if not os.path.exists(os.path.dirname(install_dst)):
@@ -391,7 +393,9 @@ def cmake(config, relative_path):
         os.makedirs(cmake_build_dir)
 
     install_prefix = (
-        os.path.join(cwd, "nvfuser") if not config.install_dir else config.install_dir
+        os.path.join(cwd, "nvfuser_common")
+        if not config.install_dir
+        else config.install_dir
     )
 
     from tools.gen_nvfuser_version import (
@@ -515,7 +519,7 @@ def run(config, version_tag, relative_path):
         # NOTE: package include files for cmake
         # TODO(crcrpar): Better avoid hardcoding `libnvfuser_codegen.so`
         # might can be treated by using `exclude_package_data`.
-        nvfuser_package_data = [
+        nvfuser_common_package_data = [
             "lib/libnvfuser_codegen.so",
             "include/nvfuser/*.h",
             "include/nvfuser/struct.inl",
@@ -556,7 +560,7 @@ def run(config, version_tag, relative_path):
                 "clean": create_clean(relative_path),
             },
             package_data={
-                "nvfuser": nvfuser_package_data,
+                "nvfuser_common": nvfuser_common_package_data,
             },
             install_requires=config.install_requires,
             extras_require={
